+++ src/main/kotlin/org/openrewrite/kotlin/KotlinIrTypeMapping.kt
@@ -15,12 +15,14 @@
  */
 package org.openrewrite.kotlin
 
+import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.descriptors.ClassKind
 import org.jetbrains.kotlin.descriptors.DescriptorVisibility
 import org.jetbrains.kotlin.descriptors.Modality
-import org.jetbrains.kotlin.fir.lazy.Fir2IrLazyClass
 import org.jetbrains.kotlin.ir.declarations.*
 import org.jetbrains.kotlin.ir.expressions.IrCall
+import org.jetbrains.kotlin.ir.expressions.IrConst
+import org.jetbrains.kotlin.ir.expressions.IrConstKind
 import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
 import org.jetbrains.kotlin.ir.symbols.IrClassSymbol
 import org.jetbrains.kotlin.ir.symbols.IrClassifierSymbol
@@ -236,7 +238,9 @@ class KotlinIrTypeMapping(typeCache: JavaTypeCache) : JavaTypeMapping<Any> {
                     methods = ArrayList(irClass.functions.toList().size)
                 }
                 val mt = methodDeclarationType(function)
-                methods.add(mt)
+                if (mt != null) {
+                    methods.add(mt)
+                }
             }
 
             for (function: IrFunction in irClass.functions) {
@@ -244,7 +248,9 @@ class KotlinIrTypeMapping(typeCache: JavaTypeCache) : JavaTypeMapping<Any> {
                     methods = ArrayList(irClass.functions.toList().size)
                 }
                 val mt = methodDeclarationType(function)
-                methods.add(mt)
+                if (mt != null) {
+                    methods.add(mt)
+                }
             }
 
             var interfaces: MutableList<JavaType.FullyQualified>? = null
@@ -265,7 +271,7 @@ class KotlinIrTypeMapping(typeCache: JavaTypeCache) : JavaTypeMapping<Any> {
             if (pt == null) {
                 val typeParameters: MutableList<JavaType> = ArrayList(irClass.typeParameters.size)
                 pt = JavaType.Parameterized(null, null, null)
-
+                typeCache.put(signature, pt)
                 val params = if (type is IrSimpleType) type.arguments else (type as IrClass).typeParameters
                 for (tp in params) {
                     typeParameters.add(type(tp))
@@ -298,7 +304,10 @@ class KotlinIrTypeMapping(typeCache: JavaTypeCache) : JavaTypeMapping<Any> {
         return gtv
     }
 
-    fun methodDeclarationType(function: IrFunction): JavaType.Method {
+    fun methodDeclarationType(function: IrFunction?): JavaType.Method? {
+        if (function == null) {
+            return null
+        }
         val signature = signatureBuilder.methodDeclarationSignature(function)
         val existing = typeCache.get<JavaType.Method>(signature)
         if (existing != null) {
@@ -344,6 +353,53 @@ class KotlinIrTypeMapping(typeCache: JavaTypeCache) : JavaTypeMapping<Any> {
             paramTypes, null, listAnnotations(function.annotations)
         )
         return method
+    }
+
+    fun methodInvocationType(type: Any?): JavaType.Method? {
+        if (type == null) {
+            return null
+        }
+        val signature = signatureBuilder.methodSignature(type)
+        val existing = typeCache.get<JavaType.Method>(signature)
+        if (existing != null) {
+            return existing
+        }
+        return methodInvocationType(type, signature)
+    }
+
+    fun methodInvocationType(type: Any, signature: String): JavaType.Method {
+        return JavaType.Method(
+            null,
+            0,
+            null,
+            "",
+            null,
+            null,
+            null,
+            null,
+            null)
+    }
+
+    fun primitive(type: Any?): JavaType.Primitive {
+        return when (type) {
+            is IrConst<*> -> {
+                when (type.kind) {
+                    IrConstKind.Int -> JavaType.Primitive.Int
+                    IrConstKind.Boolean -> JavaType.Primitive.Boolean
+                    IrConstKind.Byte -> JavaType.Primitive.Byte
+                    IrConstKind.Char -> JavaType.Primitive.Char
+                    IrConstKind.Double -> JavaType.Primitive.Double
+                    IrConstKind.Float -> JavaType.Primitive.Float
+                    IrConstKind.Long -> JavaType.Primitive.Long
+                    IrConstKind.Null -> JavaType.Primitive.Null
+                    IrConstKind.Short -> JavaType.Primitive.Short
+                    IrConstKind.String -> JavaType.Primitive.String
+                }
+            }
+            else -> {
+                TODO()
+            }
+        }
     }
 
     private fun typeProjection(type: Any, signature: String): JavaType {
