+++ src/main/kotlin/org/openrewrite/kotlin/internal/PsiElementAssociations.kt
@@ -15,165 +15,171 @@
  */
 package org.openrewrite.kotlin.internal
 
-import org.jetbrains.kotlin.KtFakeSourceElement
 import org.jetbrains.kotlin.KtRealPsiSourceElement
+import org.jetbrains.kotlin.com.intellij.openapi.util.TextRange
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
+import org.jetbrains.kotlin.com.intellij.psi.PsiElementVisitor
+import org.jetbrains.kotlin.com.intellij.psi.PsiFile
 import org.jetbrains.kotlin.fir.FirElement
-import org.jetbrains.kotlin.fir.declarations.FirDeclaration
-import org.jetbrains.kotlin.fir.declarations.FirFile
-import org.jetbrains.kotlin.fir.expressions.FirExpression
-import org.jetbrains.kotlin.fir.expressions.FirFunctionCall
-import org.jetbrains.kotlin.fir.expressions.FirReturnExpression
+import org.jetbrains.kotlin.fir.backend.FirMetadataSource
 import org.jetbrains.kotlin.fir.expressions.impl.FirElseIfTrueCondition
 import org.jetbrains.kotlin.fir.expressions.impl.FirSingleExpressionBlock
-import org.jetbrains.kotlin.fir.references.FirResolvedNamedReference
-import org.jetbrains.kotlin.fir.references.resolved
-import org.jetbrains.kotlin.fir.resolve.dfa.DfaInternals
-import org.jetbrains.kotlin.fir.symbols.FirBasedSymbol
-import org.jetbrains.kotlin.fir.symbols.impl.FirConstructorSymbol
-import org.jetbrains.kotlin.fir.types.FirResolvedTypeRef
-import org.jetbrains.kotlin.fir.visitors.FirDefaultVisitor
-import org.jetbrains.kotlin.psi
-import org.jetbrains.kotlin.psi.KtDeclaration
-import org.jetbrains.kotlin.psi.KtDotQualifiedExpression
+import org.jetbrains.kotlin.ir.IrElement
+import org.jetbrains.kotlin.ir.declarations.*
+import org.jetbrains.kotlin.ir.expressions.*
+import org.jetbrains.kotlin.ir.visitors.IrElementVisitor
+import org.jetbrains.kotlin.psi.KtClass
+import org.jetbrains.kotlin.psi.KtConstantExpression
 import org.jetbrains.kotlin.psi.KtExpression
+import org.jetbrains.kotlin.psi.KtProperty
+import org.jetbrains.kotlin.psi.KtStringTemplateExpression
 import org.openrewrite.java.tree.JavaType
-import org.openrewrite.kotlin.KotlinTypeMapping
+import org.openrewrite.kotlin.KotlinIrTypeMapping
 
-class PsiElementAssociations(val typeMapping: KotlinTypeMapping, val file: FirFile) {
+class PsiElementAssociations(val typeMapping: KotlinIrTypeMapping, private val psiFile: PsiFile, val file: IrFile) {
 
-    private val elementMap: MutableMap<PsiElement, MutableList<FirInfo>> = HashMap()
+    private val psiMap: MutableMap<TextRange, MutableSet<PsiElement>> = HashMap()
+    private val elementMap: MutableMap<PsiElement, MutableSet<IrInfo>> = HashMap()
 
     fun initialize() {
         var depth = 0
-        object : FirDefaultVisitor<Unit, MutableMap<PsiElement, MutableList<FirInfo>>>() {
-            override fun visitElement(element: FirElement, data: MutableMap<PsiElement, MutableList<FirInfo>>) {
-                if (element.source != null && element.source.psi != null) {
-                    val psiElement = element.source!!.psi!!
-                    val firInfo = FirInfo(element, depth)
-                    data.computeIfAbsent(psiElement) { ArrayList() } += firInfo
+        object : PsiElementVisitor() {
+            override fun visitElement(element: PsiElement) {
+                psiMap.computeIfAbsent(element.textRange) { HashSet() } += element
+                element.acceptChildren(this)
+            }
+        }.visitFile(psiFile)
+
+        object : IrElementVisitor<Unit, MutableMap<PsiElement, MutableSet<IrInfo>>> {
+            override fun visitElement(element: IrElement, data: MutableMap<PsiElement, MutableSet<IrInfo>>) {
+                if (element is IrMetadataSourceOwner) {
+                    if (element.metadata is FirMetadataSource) {
+                        when (element.metadata) {
+                            is FirMetadataSource.File -> {
+                                val source = (element.metadata!! as FirMetadataSource.File).files[0].source
+                                if (source is KtRealPsiSourceElement) {
+                                    elementMap.computeIfAbsent(source.psi) { HashSet() } += IrInfo(element, depth)
+                                }
+                            }
+
+                            is FirMetadataSource.Class -> {
+                                val source = (element.metadata!! as FirMetadataSource.Class).fir.source
+                                if (source is KtRealPsiSourceElement) {
+                                    elementMap.computeIfAbsent(source.psi) { HashSet() } += IrInfo(element, depth)
+                                }
+                            }
+
+                            is FirMetadataSource.Function -> {
+                                val source = (element.metadata!! as FirMetadataSource.Function).fir.source
+                                if (source is KtRealPsiSourceElement) {
+                                    elementMap.computeIfAbsent(source.psi) { HashSet() } += IrInfo(element, depth)
+                                }
+                            }
+
+                            is FirMetadataSource.Property -> {
+                                val source = (element.metadata!! as FirMetadataSource.Property).fir.source
+                                if (source is KtRealPsiSourceElement) {
+                                    elementMap.computeIfAbsent(source.psi) { HashSet() } += IrInfo(element, depth)
+                                }
+                            }
+
+                            is FirMetadataSource.Script -> {
+                                val source = (element.metadata!! as FirMetadataSource.Script).fir.source
+                                if (source is KtRealPsiSourceElement) {
+                                    elementMap.computeIfAbsent(source.psi) { HashSet() } += IrInfo(element, depth)
+                                }
+                            }
+                        }
+                    }
+                } else if (element.startOffset >= 0 && element.endOffset >= 0) {
+                    val textRange = TextRange.create(element.startOffset, element.endOffset)
+                    if (psiMap.containsKey(textRange)) {
+                        val psi = psiMap[textRange]
+                        when (element) {
+                            is IrConst<*> -> {
+                                for (p in psi!!.toList()) {
+                                    if (p is KtConstantExpression || p is KtStringTemplateExpression) {
+                                        elementMap.computeIfAbsent(p) { HashSet() } += IrInfo(element, depth)
+                                        break
+                                    }
+                                }
+                            }
+                            is IrBlockBody, is IrExpressionBody, is IrWhen, is IrBranch -> {
+                                // Skip
+                            }
+                            else -> {
+                                // TODO: map if necessary
+                            }
+                        }
+                    }
                 }
+
                 depth++
                 element.acceptChildren(this, data)
-                if (element is FirResolvedTypeRef) {
-                    // not sure why this isn't taken care of by `FirResolvedTypeRefImpl#acceptChildren()`
-                    element.delegatedTypeRef?.accept(this, data)
-                }
                 depth--
             }
         }.visitFile(file, elementMap)
-
         validate()
     }
 
     private fun validate() {
-        var found1ToNMapping = false
-        elementMap.forEach { (psi, firList) ->
-            var fakeCount = 0
-            var realCount = 0
-            var otherCount = 0
-            for (firElement in firList) {
-                if (firElement.fir.source is KtRealPsiSourceElement) {
-                    realCount++
-                } else if (firElement.fir.source is KtFakeSourceElement) {
-                    fakeCount++
-                } else {
-                    otherCount++
-                }
-            }
-
-            // print out logs, debug purpose only, to be removed after complete parser
-            if (false) {
-                found1ToNMapping = realCount > 1
-
-                println("---------")
-                println("PSI: $psi")
-                println("FIR: $firList")
-
-                println("Found 1 to $realCount Real mapping!")
-                println("    types from $realCount Real elements:")
-                var firstUnknown = false
-                var hasNonUnknown = false
-                for ((index, firElement) in firList.withIndex()) {
-                    if (firElement.fir.source is KtRealPsiSourceElement) {
-                        val type = typeMapping.type(firElement.fir).toString()
-                        if (index == 0 && type.equals("Unknown")) {
-                            firstUnknown = true
-                        }
-
-                        if (!type.equals("Unknown")) {
-                            hasNonUnknown = true
-                        }
-
-                        val padded = "        -$type".padEnd(30, ' ')
-                        println("$padded - $firElement")
-                    }
-                }
-
-                if (firstUnknown && hasNonUnknown) {
-                    throw IllegalArgumentException("First type is Unknown!")
-                }
-
-                println("    types from $fakeCount Fake elements:")
-                for (firElement in firList) {
-                    if (firElement.fir.source is KtFakeSourceElement) {
-                        val type = typeMapping.type(firElement.fir).toString()
-                        val padded = "        -$type".padEnd(30, ' ')
-                        println("$padded - $firElement")
-
-                    }
-                }
-            }
-        }
-
-        if (found1ToNMapping) {
-            // throw IllegalArgumentException("Found 1 to N real mapping!")
-        }
+        // TODO
     }
 
-    fun type(psiElement: PsiElement, ownerFallBack: FirBasedSymbol<*>?): JavaType? {
-        val fir = primary(psiElement)
-        return if (fir != null) typeMapping.type(fir, ownerFallBack) else null
-    }
-
-    fun symbol(psi: KtDeclaration?): FirBasedSymbol<*>? {
-        val fir = fir(psi) { it is FirDeclaration }
-        return if (fir != null) (fir as FirDeclaration).symbol else null
-    }
-
-    fun symbol(psi: KtExpression?): FirBasedSymbol<*>? {
-        val fir = fir(psi) { it is FirResolvedNamedReference }
-        return if (fir is FirResolvedNamedReference) fir.resolvedSymbol else null
+    fun type(psiElement: PsiElement): JavaType? {
+        val ir = primary(psiElement)
+        return if (ir != null) typeMapping.type(ir) else null
     }
 
     fun primary(psiElement: PsiElement) =
-        fir(psiElement) { it.source is KtRealPsiSourceElement }
+        ir(psiElement) {
+            when (psiElement) {
+                is KtClass -> {
+                    it is IrClass
+                }
 
-    fun component(psiElement: PsiElement) =
-        fir(psiElement) { it is FirFunctionCall}
+                is KtConstantExpression, is KtStringTemplateExpression -> {
+                    it is IrConst<*>
+                }
 
-    fun fir(psi: PsiElement?, filter: (FirElement) -> Boolean) : FirElement? {
+                is KtProperty -> {
+                    it is IrProperty
+                }
+
+                else -> {
+                    it is IrElement
+                }
+            }
+        }
+
+    fun ir(psi: PsiElement?, filter: (IrElement) -> Boolean): IrElement? {
         var p = psi
         while (p != null && !elementMap.containsKey(p)) {
             p = p.parent
-            // don't skip KtDotQualifiedExpression for field access
-//            if (p is KtDotQualifiedExpression) {
-//                return null
-//            }
         }
 
         if (p == null) {
             return null
         }
 
-        val allFirInfos = elementMap[p]!!
-        val directFirInfos = allFirInfos.filter { filter.invoke(it.fir) }
-        return if (directFirInfos.isNotEmpty())
-            directFirInfos[0].fir
-        else if (allFirInfos.isNotEmpty())
-            allFirInfos[0].fir
+        val allIrInfos = elementMap[p]!!
+        val directIrInfos = allIrInfos.filter { filter.invoke(it.ir) }
+        return if (directIrInfos.isNotEmpty())
+            directIrInfos[0].ir
+        else if (directIrInfos.isNotEmpty())
+            directIrInfos[0].ir
         else
             null
+    }
+
+    fun methodDeclarationType(psiElement: PsiElement): IrFunction? {
+        val ir = primary(psiElement)
+        return if (ir is IrFunction) ir else null
+    }
+
+    fun methodInvocationType(psiElement: PsiElement): IrElement? {
+        val ir = primary(psiElement)
+        return if (ir is IrFunction) ir else null
     }
 
     enum class ExpressionType {
@@ -183,32 +189,13 @@ class PsiElementAssociations(val typeMapping: KotlinTypeMapping, val file: FirFi
     }
 
     fun getFunctionType(psi: KtExpression): ExpressionType? {
-        val fir = fir(psi) { it is FirFunctionCall } as? FirFunctionCall
+        val ir = ir(psi) { it is IrCall } as? IrCall
 
-        if (fir == null) {
+        if (ir == null) {
             return null
         }
 
-        return if (fir.calleeReference.resolved != null) {
-            when (fir.calleeReference.resolved!!.resolvedSymbol) {
-                is FirConstructorSymbol -> ExpressionType.CONSTRUCTOR
-                else -> ExpressionType.METHOD_INVOCATION
-            }
-        } else {
-            return null
-            // throw UnsupportedOperationException("Null resolved symbol on FirFunctionCall: $psi")
-        }
-    }
-
-    @OptIn(DfaInternals::class)
-    fun getExpressionType(psi: KtExpression): ExpressionType? {
-        val fir = fir(psi) { it is FirExpression }
-        return if (fir is FirReturnExpression) {
-            ExpressionType.RETURN_EXPRESSION
-        } else {
-            // TODO, other expression type if needed
-            null
-        }
+        return null
     }
 
     private fun PsiElement.customToString(): String {
@@ -217,31 +204,31 @@ class PsiElementAssociations(val typeMapping: KotlinTypeMapping, val file: FirFi
 
     override fun toString(): String {
         val sb = StringBuilder()
-        elementMap.forEach{ (psi, firs) ->
-            sb.append(psi.customToString()).append("\n")
-            firs.forEach{ fir ->
-                sb.append("  - $fir\n")
-            }
-            sb.append("\n")
-        }
+//        elementMap.forEach{ (psi, firs) ->
+//            sb.append(t).append("\n")
+//            firs.forEach{ fir ->
+//                sb.append("  - $fir\n")
+//            }
+//            sb.append("\n")
+//        }
         return sb.toString()
     }
 
-    private class FirInfo(
-        val fir: FirElement,
+    private class IrInfo(
+        val ir: IrElement,
         val depth: Int,
     ) {
         override fun toString(): String {
-            val s = PsiTreePrinter.printFirElement(fir)
+            val s = PsiTreePrinter.printIrElement(ir)
             return "FIR($depth, $s)"
         }
     }
 
     companion object {
-        fun printElement(firElement: FirElement) : String {
+        fun printElement(firElement: FirElement): String {
             if (firElement is FirSingleExpressionBlock) {
                 return PsiTreePrinter.firElementToString(firElement.statement)
-            } else  if (firElement is FirElseIfTrueCondition) {
+            } else if (firElement is FirElseIfTrueCondition) {
                 return "true";
             }
